% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-config.R
\name{get_token}
\alias{get_token}
\title{Getting the token}
\usage{
get_token()
}
\value{
String.
}
\description{
Function to retrieve user token.
}
\details{
Using \code{get_token} allows you to retrieve your API token.
}
