% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/govquery.R
\name{gov_query}
\alias{gov_query}
\title{Download Data from data.gov.gr API}
\usage{
gov_query(
  dataset = NULL,
  start_date = NULL,
  end_date = NULL,
  gov_token = get_token()
)
}
\arguments{
\item{dataset}{dataset to grab data from e.g. mdg_emvolio for vaccination statistics}

\item{start_date}{starting date to grab data from}

\item{end_date}{ending date to grab data from}

\item{gov_token}{data.gov.gr token}
}
\value{
The result of the query in a dataframe format.
}
\description{
The GET function for the request
}
\examples{
\dontrun{
gov_query("mdg_emvolio","2022-07-04","2022-07-11")}

}
