% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-config.R
\name{set_token}
\alias{set_token}
\title{Setting the token}
\usage{
set_token(token)
}
\arguments{
\item{token}{data.gov.gr API token}
}
\value{
None
}
\description{
Function to set the default token for use with data.gov.gr API.
}
\details{
Using \code{set_token} allows you to set your API token once and it is then used for
all functions that access the API.
See https://www.data.gov.gr/token to get your API token.
}
\examples{
set_token("token abcd12344321")

}
