% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{ua}
\alias{ua}
\title{Setting User Agent}
\format{
An object of class \code{request} of length 7.
}
\usage{
ua
}
\description{
The user agent to include with the call to the API.
}
\keyword{datasets}
